VC compiler warning fixes

+++ emboss/trimseq.c
@@ -226,7 +226,7 @@ static ajint trimseq_trim(const AjPSeq seq,
 	}
 
 	/* what is the percentage of bad characters in this window */
-	pc = 100.0 * (float)count/(float)window;
+	pc = (float)100.0 * (float)count/(float)window;
 	/* Need to trim this window? */
 	if(pc < percent)
 	    break;
